chore(build): update build configuration and dependencies
Update the build configuration by changing the recipe names in `stringutils.yml` and `rewrite.yml` to use the new package structure. Remove the OpenRewrite build publish and nebula release plugins, and instead use the java-library and maven-publish plugins. Add GPG signing configuration for publications and update the project information in the POM.